requirements.txt
# Django
Django>=4.2,<5.0

# Django REST Framework
djangorestframework>=3.14.0

# PostgreSQL Driver
psycopg2-binary>=2.9.0

# Google OAuth & JWT Authentication (SSO 登录必需)
google-auth>=2.16.0
djangorestframework-simplejwt>=5.2.0
requests>=2.31.0

# CORS (如果前端需要跨域请求)
# django-cors-headers>=4.0.0

# 环境变量管理 (推荐使用)
# python-decouple>=3.8

# 开发工具
# django-debug-toolbar>=4.0.0